+++ packages/netlify-cms-lib-auth/src/implicit-oauth.js
@@ -1,7 +1,7 @@
 import { Map } from 'immutable';
 import trim from 'lodash/trim';
 import trimEnd from 'lodash/trimEnd';
-import { createNonce, validateNonce, isInsecureProtocol } from './util';
+import { createNonce, validateNonce, isInsecureProtocol } from './utils';
 
 export default class ImplicitAuthenticator {
   constructor(config = {}) {
