+++ packages/netlify-cms-lib-auth/src/pkce-oauth.js
@@ -1,6 +1,6 @@
 import trim from 'lodash/trim';
 import trimEnd from 'lodash/trimEnd';
-import { createNonce, validateNonce, isInsecureProtocol } from './util';
+import { createNonce, validateNonce, isInsecureProtocol } from './utils';
 
 async function sha256(text) {
   return Array.from(
