+++ packages/netlify-cms-lib-auth/src/utils.js
